rearranged index and form pages
